More layout changes - Created LayoutBuilder service and used in AppComponent - Added LayoutDirectivesModule to the imports and exports of every main layout module - Removed LayoutDirectivesModule from AppModule imports as it was un-needed
