+++ projects/app/src/app/app.module.ts
@@ -4,10 +4,7 @@ import { NgModule, Provider } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-// import { FootersModule } from './features/footers';
-// import { HeadersModule } from './features/headers';
-// import { SectionsModule } from './features/layouts';
-import { LayoutDirectivesModule, LayoutTemplatesModule } from './features';
+import { LayoutTemplatesModule } from './features';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
@@ -24,12 +21,7 @@ if (environment.local) {
     BrowserModule,
     BrowserAnimationsModule,
     HttpClientModule,
-
-    LayoutDirectivesModule,
     LayoutTemplatesModule
-    // FootersModule,
-    // HeadersModule,
-    // SectionsModule
   ],
   declarations: [AppComponent],
   providers: [...dynamicProviders],
