+++ projects/app/src/app/features/layouts/footers/footers.module.ts
@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
+import { LayoutDirectivesModule } from '../directives';
 import { FooterDefaultComponent } from './components';
 
 const exportedDeclarations: any[] = [
@@ -8,8 +9,8 @@ const exportedDeclarations: any[] = [
 ];
 
 @NgModule({
-  imports: [CommonModule],
+  imports: [CommonModule, LayoutDirectivesModule],
   declarations: exportedDeclarations,
-  exports: exportedDeclarations
+  exports: [...exportedDeclarations, LayoutDirectivesModule]
 })
 export class LayoutFootersModule {}
