+++ projects/app/src/app/features/layouts/headers/headers.module.ts
@@ -4,6 +4,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatToolbarModule } from '@angular/material/toolbar';
 
+import { LayoutDirectivesModule } from '../directives';
 import { HeaderDefaultComponent } from './components';
 
 const exportedDeclarations: any[] = [
@@ -11,8 +12,8 @@ const exportedDeclarations: any[] = [
 ];
 
 @NgModule({
-  imports: [CommonModule, MatButtonModule, MatIconModule, MatToolbarModule],
+  imports: [CommonModule, LayoutDirectivesModule, MatButtonModule, MatIconModule, MatToolbarModule],
   declarations: exportedDeclarations,
-  exports: exportedDeclarations
+  exports: [...exportedDeclarations, LayoutDirectivesModule]
 })
 export class LayoutHeadersModule {}
