+++ projects/app/src/app/features/layouts/index.ts
@@ -3,4 +3,5 @@ export * from './footers';
 export * from './headers';
 export * from './models';
 export * from './sections';
+export * from './services';
 export * from './templates';
