+++ projects/app/src/app/features/layouts/sections/sections.module.ts
@@ -6,6 +6,7 @@ import { MatListModule } from '@angular/material/list';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatToolbarModule } from '@angular/material/toolbar';
 
+import { LayoutDirectivesModule } from '../directives';
 import { SectionDefaultComponent } from './components';
 
 const exportedDeclarations: any[] = [
@@ -13,8 +14,8 @@ const exportedDeclarations: any[] = [
 ];
 
 @NgModule({
-  imports: [CommonModule, MatButtonModule, MatIconModule, MatListModule, MatSidenavModule, MatToolbarModule],
+  imports: [CommonModule, LayoutDirectivesModule, MatButtonModule, MatIconModule, MatListModule, MatSidenavModule, MatToolbarModule],
   declarations: exportedDeclarations,
-  exports: exportedDeclarations
+  exports: [...exportedDeclarations, LayoutDirectivesModule]
 })
 export class LayoutSectionsModule {}
