+++ projects/app/src/app/features/layouts/services/index.ts
@@ -1,0 +1,1 @@
+export * from './layout-builder.service';
